dicom-images-backend/requirements.txt
Django==4.1
djangorestframework==3.14.0
mysqlclient==2.0.1
django-mysql==3.9
PyJWT==1.7.1
django-cors-headers==3.5.0
pydicom
pillow
numpy
graphene-django